requirements.txt
alabaster==0.7.10
anaconda-client==1.2.2
anaconda-project==0.8.2
apns2==0.4.1
asn1crypto==0.24.0
blaze==0.10.1
certifi==2018.10.15
cffi==1.12.3
chardet==3.0.4
Click==7.0
clyent==1.2.1
cryptography==2.6.1
datashape==0.5.2
decorator==4.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.20
django-allauth==0.39.1
django-push-notifications==1.6.0
django-rest-auth==0.9.5
django-tastypie==0.14.2
django-uuidfield==0.5.0
djangorestframework==3.9.2
Flask==1.0.2
Flask-Cors==3.0.7
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multipledispatch==0.6.0
networkx==2.3
numpy==1.16.3
oauthlib==3.0.1
odo==0.5.0
pandas==0.24.2
psutil==5.6.1
py-vapid==1.5.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2017.3
pywebpush==1.9.3
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
ruamel.yaml==0.15.94
six==1.11.0
SQLAlchemy==1.2.1
static3==0.7.0
toolz==0.9.0
tornado==6.0.2
urllib3==1.22
Werkzeug==0.15.2
